adds lib for functions
